# db: reduce comparisons in mergingIterHeap